+++ react/ProductSummarySKUSelector.tsx
@@ -217,6 +217,7 @@ function ProductSummarySKUSelector(props: Props) {
       <SKUSelector
         onSKUSelected={handleSKUSelected}
         {...props}
+        variations={props.visibleVariations}
         classes={skuSelectorClasses}
       />
     </div>