+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- I18n IT, FR, KO and NL translations
+
+### Changed
+- Crowdin configuration file
 
 ## [2.71.3] - 2021-05-03